网络读取线程都归并至ThreadReadVritualWebDataz中了，只有网易日线读取线程需要加以修改，目前尚未归并。清除废弃的函数定义。

+++ GoogleUnitTest/MockWorldStockTest.cpp
@@ -51,6 +51,6 @@ namespace StockAnalysisTest {
 		CMockWorldStock stock;
 		EXPECT_CALL(stock, UpdateEPSSurpriseDB)
 			.Times(1);
-		EXPECT_EQ(ThreadUpdateEPSSurpriseDB(&stock), 41);
+		EXPECT_EQ(ThreadUpdateEPSSurpriseDB(&stock), (UINT)41);
 	}
 }
